@@ -1,22 +1,36 @@
 name: dev_yahia
 description: Portfolio for DevYahia
 
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
-version: 1.0.0+1
+version: 2.0.0+2
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  
-  flutter_icons: ^1.1.0
-  cupertino_icons: ^1.0.2
-  url_launcher: ^6.0.3
-  url_launcher_web: ^2.0.0
+  # icons
+  community_material_icon: ^5.9.55
+  cupertino_icons: ^1.0.3
+  # links
+  url_launcher: ^6.0.10
+  url_launcher_web: ^2.0.4
   mailto: ^2.0.0
+  # image
+  image_picker: ^0.8.4
+  # misc
+  fluttertoast: ^8.0.8
+  url_strategy: ^0.2.0
+  universal_html: ^2.0.8
+  # firebase
+  firebase_core: ^1.6.0
+  cloud_firestore: ^2.5.2
+  firebase_storage: ^10.0.3
+  # state
+  get: ^4.3.8
+  google_fonts: ^2.1.0
 
 dev_dependencies:
   flutter_test:
