@@ -4,11 +4,9 @@ class AppIcon extends StatelessWidget {
   final String path;
   final double size;
   final Color color;
-  final Color background;
+  final Color? background;
 
-  const AppIcon(this.path,
-      {Key key, this.size = 25, this.color = Colors.white, this.background})
-      : super(key: key);
+  const AppIcon(this.path, {Key? key, this.size = 25, this.color = Colors.white, this.background}) : super(key: key);
   @override
   Widget build(BuildContext context) {
     return Container(
