@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 mixin AppColors {
-  static final yellow = Colors.deepPurple[800];
+  static final yellow = Colors.red[800];
 
   static final black = Colors.black87;
 
