@@ -1,23 +1,11 @@
+import 'package:dev_yahia/my_app.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
+import 'package:url_strategy/url_strategy.dart';
 
-import 'ui/home.dart';
-import 'config/colors.dart';
-
-void main() {
+Future<void> main() async {
+  setPathUrlStrategy();
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp();
   runApp(MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'DevYahia',
-      debugShowCheckedModeBanner: false,
-      theme: ThemeData(
-        accentColor: AppColors.yellow,
-        visualDensity: VisualDensity.adaptivePlatformDensity,
-      ),
-      home: Home(),
-    );
-  }
 }
